refactor: Update code according to IntelliJ analyze suggestions (#1594)
* refactor: Update code according to IntelliJ analyze suggestions ("Artemis mini")

* More fixes according to IntelliJ analyze suggestions ("Artemis"); mostly restrictions in visibility

* fix: fix up some review comments

---------

+++ common/src/main/java/com/wynntils/utils/wynn/LocationUtils.java
@@ -45,8 +45,7 @@ public final class LocationUtils {
     }
 
     public static void shareCompass(String target, Location compass) {
-        String locationString =
-                "My compass is at [" + (int) compass.x + ", " + (int) compass.y + ", " + (int) compass.z + "]";
+        String locationString = "My compass is at [" + compass.x + ", " + compass.y + ", " + compass.z + "]";
 
         LocationUtils.sendShareMessage(target, locationString);
     }
